storage: make use of busy timeouts and remove lock
We used the mutex ensure there are no attempts for parallel write
transactons. Such attempts resulted in errors and affected the overall
stability of the node. Now the storage busy timeout is used. It allows
to wait for another write transaction to complete instead of failing
immediately. This allows us to remove the mutex.